refactor: decoupled demos from dynamic_programming
