+++ CMakeLists.txt
@@ -46,6 +46,7 @@ set(SRC_DIRS
     demos/advanced_graph_algorithms
     demos/string_algorithms
     demos/advanced_sorting_algorithms
+    demos/dynamic_programming
     features/telemetry
     features/serialization
     features/dp_visualizer
